refactor: Remove duplicated colors.rs file (#11990)
