+++ cli/Cargo.toml
@@ -88,7 +88,6 @@ serde = { version = "1.0.129", features = ["derive"] }
 shell-escape = "0.1.5"
 sourcemap = "6.0.1"
 tempfile = "3.2.0"
-termcolor = "1.1.2"
 text-size = "1.1.0"
 tokio = { version = "1.10.1", features = ["full"] }
 tokio-rustls = "0.22.0"
