+++ cli/diagnostics.rs
@@ -1,6 +1,6 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-use crate::colors;
+use deno_runtime::colors;
 
 use deno_core::serde::Deserialize;
 use deno_core::serde::Deserializer;
@@ -421,9 +421,9 @@ impl Error for Diagnostics {}
 #[cfg(test)]
 mod tests {
   use super::*;
-  use colors::strip_ansi_codes;
   use deno_core::serde_json;
   use deno_core::serde_json::json;
+  use test_util::strip_ansi_codes;
 
   #[test]
   fn test_de_diagnostics() {
