+++ cli/diff.rs
@@ -215,7 +215,7 @@ mod tests {
 
   fn run_test(diff_text1: &str, diff_text2: &str, expected_output: &str) {
     assert_eq!(
-      colors::strip_ansi_codes(&diff(diff_text1, diff_text2,)),
+      test_util::strip_ansi_codes(&diff(diff_text1, diff_text2,)),
       expected_output,
     );
   }
